ongoing

+++ solarpoolheater/PumpControl.h
@@ -5,6 +5,6 @@ void setupPumpControl();
 void tickPumpControl();
 
 // the cumulative time that the pump has been running, in seconds.  resets at midnight.
-extern unsigned long pumpRuntimeSeconds;
+extern float pumpRuntimeSeconds;
 
 #endif
